feat [SRTP-724] refactoring performance test folder
